feature/13/movie-hover > on hover, basic info displays with a functional like button
