+++ src/assets/styles/_buttons.scss
@@ -5,7 +5,7 @@
 }
 
 .button__upvote {
-	display: none;
+	// waiting on design
 }
 
 .button__carousel {
