+++ src/assets/styles/_variables.scss
@@ -1,2 +1,6 @@
 $red: #ee2024;
 $gray: #2b2b2b;
+
+// To more easily manage our z-indexes we're defining them one by one.
+$z-carousel-image: 3;
+$z-carousel-info: 2;
