+++ src/assets/styles/styles.scss
@@ -9,3 +9,7 @@ main {
 	margin: 0 auto;
 	position: relative;
 }
+
+.hidden {
+	display: none;
+}
