+++ src/components/ListMovie.js
@@ -8,38 +8,7 @@ class ListMovie extends React.Component {
 		super(props);
 
 		this.movies = Object.keys(this.props.movies);
-
-		this.upVote = this.upVote.bind(this);
 		this.firebaseRef = firebaseConfig.database.ref('/movies');
-
-		console.log(this.movies, ' this.movies')
-	}
-
-	upVote (selectedMovie, index) {
-		const upvotesRef = Firebase.database().ref('movies/' + selectedMovie.details.key);
-
-		if(Firebase.auth().currentUser != null) {
-			function toggleVote(firebaseRef, uid) {
-				firebaseRef.transaction((post) => {
-					if (post) {
-						if(post.stars && post.stars[uid]) {
-							post.likes--;
-							post.stars[uid] = null;
-						}else{
-							post.likes++;
-							if (!post.stars) {
-					          post.stars = {};
-					        }
-					        post.stars[uid] = true;
-						}
-					}
-					return post;
-				});
-			}
-
-			toggleVote(upvotesRef, Firebase.auth().currentUser.uid);
-
-		}
 	}
 
 	render () {
@@ -54,9 +23,9 @@ class ListMovie extends React.Component {
 								key={key} 
 								index={key} 
 								movies={this.props.movies} 
-								details={this.props.movies[key]} 
+								movie={this.props.movies[key]} 
 								removeMovie={this.props.removeMovie} 
-								upVote={this.upVote} 
+								upvoteMovie={this.props.upvoteMovie} 
 							/>
 						)
 					}
@@ -68,17 +37,17 @@ class ListMovie extends React.Component {
 };
 
 function MovieAppList (props) {
-	const { details, index } = props;
-	const UpvoteButton = <button onClick={() => props.upVote({details}, index)}>+1</button>;
-	const RemoveButton = <button onClick={() => props.removeMovie(index)}>&times;</button>;
+	const { movie, index } = props;
+	const UpvoteButton = <button className="button__upvote" onClick={() => props.upvoteMovie({movie}, index)}>+1</button>;
+	const RemoveButton = <button className="button__delete" onClick={() => props.removeMovie(index)}>&times;</button>;
 
 	return (
 		<li>
-			<span className="yellow">{ details.name }</span>
-			<span className="MovieApp__result-list-likes">{ details.likes }</span>
+			<span className="yellow">{ movie.name }</span>
+			<span className="MovieApp__result-list-likes">{ movie.likes }</span>
 			<span className="green"></span>
 			<span className="red">{ UpvoteButton }</span>
-			<span className="button__delete blue">{ RemoveButton }</span>
+			<span className="blue hidden">{ RemoveButton }</span>
 		</li>
 	)
 }
